requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
matplotlib==3.7.1
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
patool==1.12
pygame==2.4.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.2
zipp==3.15.0
